Add checking admin user.

+++ app/controllers/users_controller.rb
@@ -1,7 +1,8 @@
 class UsersController < ApplicationController
   before_action :set_user, only: %i[edit update show destroy]
   before_action :require_user, only: %i[edit update destroy]
-  before_action :require_same_user, only: %i[edit update destroy]
+  before_action :require_same_user, only: %i[edit update]
+  before_action :require_same_or_admin_user, only: :destroy
 
   def index
     @users = User.paginate(page: params[:page], per_page: 5)
@@ -39,7 +40,7 @@ class UsersController < ApplicationController
 
   def destroy
     @user.destroy
-    session.delete(:user_id)
+    session.delete(:user_id) if @user == current_user
     flash[:success] = 'You have deleted your account with articles'
     redirect_to root_path
   end
@@ -57,6 +58,13 @@ class UsersController < ApplicationController
   def require_same_user
     if current_user != @user
       flash[:warning] = 'You can only edit your own account!'
+      redirect_to @user
+    end
+  end
+
+  def require_same_or_admin_user
+    if current_user != @user && !current_user.admin?
+      flash[:warning] = 'You can only delete your own account!'
       redirect_to @user
     end
   end
